requirements.txt
# Dependências do Sistema de Seguros
# Sprint 4 - Persistência Híbrida (MySQL + MongoDB)

# Gerenciamento de variáveis de ambiente
python-dotenv==1.0.0

# MySQL
mysql-connector-python==8.2.0

# MongoDB
pymongo==4.6.0

# Testes (Sprint 4)
pytest==7.4.3
pytest-cov==4.1.0

# Padronização de Código (Sprint 4)
black==23.12.1
ruff==0.14.2
isort==5.13.2
